fix: x64 — clean pointer and size conversions

+++ code/src/botlib/be_aas_bspq3.c
@@ -422,14 +422,14 @@ void AAS_ParseBSPEntities( void ) {
 				return;
 			} //end if
 			StripDoubleQuotes( token.string );
-			bufsize += strlen( token.string ) + 1;
+			bufsize += (int)strlen( token.string ) + 1;
 			if ( !PS_ExpectTokenType( script, TT_STRING, 0, &token ) ) {
 				AAS_FreeBSPEntities();
 				FreeScript( script );
 				return;
 			} //end if
 			StripDoubleQuotes( token.string );
-			bufsize += strlen( token.string ) + 1;
+			bufsize += (int)strlen( token.string ) + 1;
 		} //end while
 		if ( strcmp( token.string, "}" ) ) {
 			ScriptError( script, "missing }\n" );
@@ -516,7 +516,7 @@ void AAS_DumpBSPData( void ) {
 //===========================================================================
 int AAS_LoadBSPFile( void ) {
 	AAS_DumpBSPData();
-	bspworld.entdatasize = strlen( botimport.BSPEntityData() ) + 1;
+	bspworld.entdatasize = (int)strlen( botimport.BSPEntityData() ) + 1;
 	bspworld.dentdata = (char *) GetClearedHunkMemory( bspworld.entdatasize );
 	memcpy( bspworld.dentdata, botimport.BSPEntityData(), bspworld.entdatasize );
 	AAS_ParseBSPEntities();
