+++ code/src/renderer/tr_image.c
@@ -1885,7 +1885,7 @@ static int hackSize;
 void term_destination( j_compress_ptr cinfo ) {
 	my_dest_ptr dest = (my_dest_ptr) cinfo->dest;
 	size_t datacount = dest->size - dest->pub.free_in_buffer;
-	hackSize = datacount;
+	hackSize = (int)datacount;
 }
 
 
@@ -2040,7 +2040,7 @@ void R_LoadImage( const char *name, byte **pic, int *width, int *height ) {
 	*width = 0;
 	*height = 0;
 
-	len = strlen( name );
+	len = (int)strlen( name );
 	if ( len < 5 ) {
 		return;
 	}
@@ -2050,7 +2050,7 @@ void R_LoadImage( const char *name, byte **pic, int *width, int *height ) {
 		if ( !*pic ) {                              //
 			char altname[MAX_QPATH];                    // try jpg in place of tga
 			strcpy( altname, name );
-			len = strlen( altname );
+			len = (int)strlen( altname );
 			altname[len - 3] = 'j';
 			altname[len - 2] = 'p';
 			altname[len - 1] = 'g';
