+++ code/src/splines/util_str.h
@@ -188,7 +188,7 @@ inline idStr::idStr
 	assert( text );
 
 	if ( text ) {
-		len = strlen( text );
+		len = (int)strlen( text );
 		EnsureAlloced( len + 1 );
 		strcpy( m_data->data, text );
 		m_data->len = len;
@@ -260,7 +260,7 @@ inline idStr::idStr
 	int len;
 
 	sprintf( text, "%.3f", num );
-	len = strlen( text );
+	len = (int)strlen( text );
 	EnsureAlloced( len + 1 );
 	strcpy( m_data->data, text );
 	m_data->len = len;
@@ -274,7 +274,7 @@ inline idStr::idStr
 	int len;
 
 	sprintf( text, "%d", num );
-	len = strlen( text );
+	len = (int)strlen( text );
 	EnsureAlloced( len + 1 );
 	strcpy( m_data->data, text );
 	m_data->len = len;
@@ -288,7 +288,7 @@ inline idStr::idStr
 	int len;
 
 	sprintf( text, "%u", num );
-	len = strlen( text );
+	len = (int)strlen( text );
 	EnsureAlloced( len + 1 );
 	strcpy( m_data->data, text );
 	m_data->len = len;
@@ -317,7 +317,7 @@ inline void idStr::append
 	assert( text );
 
 	if ( text ) {
-		len = length() + strlen( text );
+		len = length() + (int)strlen( text );
 		EnsureAlloced( len + 1 );
 
 		strcat( m_data->data, text );
@@ -413,7 +413,7 @@ inline void idStr::operator=
 	}
 
 	if ( !m_data ) {
-		len = strlen( text );
+		len = (int)strlen( text );
 		EnsureAlloced( len + 1, false );
 		strcpy( m_data->data, text );
 		m_data->len = len;
@@ -433,7 +433,7 @@ inline void idStr::operator=
 		// Great, we're aliasing.  We're copying from inside ourselves.
 		// This means that I don't have to ensure that anything is alloced,
 		// though I'll assert just in case.
-		int diff = text - m_data->data;
+		int diff = (int)( text - m_data->data );
 		int i;
 
 		assert( strlen( text ) < (unsigned) m_data->len );
@@ -450,7 +450,7 @@ inline void idStr::operator=
 		return;
 	}
 
-	len = strlen( text );
+	len = (int)strlen( text );
 	EnsureAlloced( len + 1, false );
 	strcpy( m_data->data, text );
 	m_data->len = len;
